chore(iam): SCRUM-78 align refresh_tokens schema with bundle
