+++ src/IAM/Infrastructure/Persistence/MySQL/Migration/Version20260114130000.php
@@ -1,0 +1,41 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Twitter\IAM\Infrastructure\Persistence\MySQL\Migration;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+final class Version20260114130000 extends AbstractMigration
+{
+    public function getDescription(): string
+    {
+        return 'Create refresh_tokens table compatible with JWT Refresh Token Bundle';
+    }
+
+    public function up(Schema $schema): void
+    {
+//        $this->addSql('DROP TABLE IF EXISTS refresh_tokens');
+
+        $this->addSql(
+            <<<SQL
+                CREATE TABLE refresh_tokens (
+                    id INT AUTO_INCREMENT NOT NULL,
+                    refresh_token VARCHAR(128) NOT NULL,
+                    username VARCHAR(255) NOT NULL,
+                    valid DATETIME NOT NULL,
+                    UNIQUE INDEX UNIQ_refresh_token (refresh_token),
+                    INDEX IDX_username (username),
+                    INDEX IDX_valid (valid),
+                    PRIMARY KEY(id)
+                ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci
+                SQL
+        );
+    }
+
+    public function down(Schema $schema): void
+    {
+        $this->addSql('DROP TABLE IF EXISTS refresh_tokens');
+    }
+}
